+++ ggc-core/src/ggc/BatchComparator.java
@@ -14,7 +14,7 @@ public class BatchComparator implements Comparator<Batch> {
             if (c.compare(_this.get_partner(), other.get_partner()) == 0) {
                 if (_this.getPrice() - other.getPrice() == 0)
                     return _this.getQuantity() - other.getQuantity();
-                return _this.getPrice() - other.getPrice();
+                return (int) _this.getPrice() - (int) other.getPrice();
             }
             return c.compare(_this.get_partner(), other.get_partner());
         }
